PageMapper bug fixed, UserProfileCreationDto bug fixed

+++ src/main/java/com/adorjanpraksa/contactsapp/web/dto/PageDto.java
@@ -10,7 +10,7 @@ public class PageDto<T> {
     private T content;
     private int pageNumber;
     private int Size;
-    private int totalElements;
+    private Long totalElements;
 
 
 }
